fixed some compilation issues and bumped cobertura version
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@1172612 13f79535-47bb-0310-9956-ffa450edef68

+++ http/src/main/java/org/apache/mina/http/HttpServerCodec.java
@@ -128,7 +128,7 @@ public class HttpServerCodec implements IoFilter {
             // send the chunk of body
             controller.callReadNextFilter(session, msg);
             // do we have reach end of body ?
-            int remaining = session.getAttribute(BODY_REMAINING_BYTES);
+            int remaining = (Integer) session.getAttribute(BODY_REMAINING_BYTES);
             remaining -= chunkSize;
 
             if (remaining <= 0) {
